Changed type of constant SIZEOF_INT32_T and SIZEOF_UINT32_T. This resolved #62

+++ trunk/cli/utils.hpp
@@ -38,10 +38,10 @@ static const std::string HISTO_FILE_TAG("histo.dat");
 static const std::string VERSION_TAG("1.0.0iqc1");
 
 // This is a constant to hold the size of an int32_t
-static const int32_t SIZEOF_INT32_T = sizeof(int32_t);
+static const size_t SIZEOF_INT32_T = sizeof(int32_t);
 
 // This is a constant to hold the size of an uint32_t
-static const int32_t SIZEOF_UINT32_T = sizeof(uint32_t);
+static const size_t SIZEOF_UINT32_T = sizeof(uint32_t);
 
 // This is a constant to hold the "zero" value for the DAS
 static const float SMALLEST_TIME_BIN = 0.01;
